Refactor: Move ExtensionContext out
